# Enable direct calls to address-taken functions from optimised functions