This change introduces another function clone with the prefix __yk_opt for address-taken functions, enabling optimised functions to call optimised versions directly.

Changes:
- Address-taken functions now get two versions: original (unoptimised with real tracing calls) and __yk_opt_ clone (optimised with dummy tracing calls)
- Non-address-taken functions keep existing behavior: original (optimised with dummy calls) and __yk_unopt_ clone (unoptimised with real calls)
- BasicBlockTracer uses metadata (__yk_opt/__yk_unopt) instead of name prefixes
- Indirect calls always target original address-taken functions to preserve function pointer semantics

This allows optimised code paths to avoid the overhead of real tracing calls when calling address-taken functions directly.


 :point_right:  Followup from this [comment](https://github.com/ykjit/ykllvm/pull/269#issuecomment-3183950657)


## Benchmarks

| Benchmark | `yklua/main` (ms) | `yklua/call-opt-functions-from-opt-context` (ms) | Ratio | Summary |
|-----------|-------------------|---------------------------------------------------|-------|---------|
| NBody/YkLuaMulti/250000 | 514 | 480 | 0.93 | 6.76% faster |
| Queens/YkLuaMulti/1000 | 474 | 444 | 0.94 | 6.14% faster |
| Sieve/YkLuaMulti/3000 | 612 | 596 | 0.97 | 2.64% faster |
| List/YkLuaMulti/1500 | 808 | 796 | 0.99 | 1.43% faster |
| Richards/YkLuaMulti/100 | 4113 | 4058 | 0.99 | 1.33% faster |
| Storage/YkLuaMulti/1000 | 3012 | 2982 | 0.99 | 0.99% faster |
| Permute/YkLuaMulti/1000 | 852 | 844 | 0.99 | 0.87% faster |
| Havlak/YkLuaMulti/1500 | 8175 | 8110 | 0.99 | 0.80% faster |
| DeltaBlue/YkLuaMulti/12000 | 1105 | 1101 | 1.00 | 0.33% faster |
| Json/YkLuaMulti/100 | 1351 | 1347 | 1.00 | 0.28% faster |
| CD/YkLuaMulti/250 | 3591 | 3583 | 1.00 | 0.21% faster |
| Mandelbrot/YkLuaMulti/500 | 128 | 129 | 1.01 | 1.06% slower |
| Bounce/YkLuaMulti/1500 | 886 | 897 | 1.01 | 1.17% slower |
| Towers/YkLuaMulti/600 | 916 | 945 | 1.03 | 3.23% slower |