@ltratt 2025-09-07
If I've understood this correctly then I *think* the following patch to yk should safely allow compiled traces to call optimised versions of functions (when they exist). Does it work for you and does it improve performance? [If it doesn't, it might be worth seeing if the "good" case is ever hit.]

```
diff --git ykrt/src/compile/jitc_yk/codegen/x64/mod.rs ykrt/src/compile/jitc_yk/codegen/x64/mod.rs
index e7da786b..451ca12a 100644
--- ykrt/src/compile/jitc_yk/codegen/x64/mod.rs
+++ ykrt/src/compile/jitc_yk/codegen/x64/mod.rs
@@ -1873,7 +1873,9 @@ impl<'a> Assemble<'a> {
                 Ok(())
             }
             x => {
-                let va = symbol_to_ptr(x).map_err(|e| CompilationError::General(e.to_string()))?;
+                let va = symbol_to_ptr(&format!("__yk_opt_{x}"))
+                    .or_else(|_| symbol_to_ptr(x))
+                    .map_err(|e| CompilationError::General(e.to_string()))?;
                 self.emit_call(iidx, fty, Some(va), None, &args)
             }
         }
```