@ltratt 2025-09-07
I quickly tried this and I couldn't see any functions `__yk_opt_X` with yklua. Perhaps I've done something wrong?